Windows screen interrupt error fix (#220)
* Windows screen interrupt error fix

Fix for argument error to win32.event.WaitForSingleObject found on Windows 10 with Python 2.7.15.

Attempting to run samples/contact_list.py without this fix results in the following:

File "C:\Python27\lib\site-packages\asciimatics\screen.py", line 2107, in wait_for_input
    rc = win32event.WaitForSingleObject(self._stdin, timeout * 1000)
SystemError: ..\Objects\longobject.c:463: bad argument to internal function

Not sure if this is the right fix, or the best fix, but the example runs after it is applied.

* Changed long cast to int cast for Python 3

Tested and working in the same manner as the initial pr commit of long casting.

+++ asciimatics/screen.py
@@ -2104,7 +2104,7 @@ if sys.platform == "win32":
 
             :param timeout: Time to wait for input in seconds (floating point).
             """
-            rc = win32event.WaitForSingleObject(self._stdin, timeout * 1000)
+            rc = win32event.WaitForSingleObject(self._stdin, int(timeout * 1000))
             if rc not in [0, 258]:
                 raise RuntimeError(rc)
